remove IJwtBlackListService and move methods to ITokenService
