--- modules/backend/src/Au5.Application/Common/Abstractions/IJwtBlacklistService.cs
@@ -1,8 +1,0 @@
-namespace Au5.Application.Common.Abstractions;
-
-public interface IJwtBlacklistService
-{
-	Task BlacklistTokenAsync(string userId, string jti, DateTime expiry);
-
-	Task<bool> IsTokenBlacklistedAsync(string userId, string jti);
-}
