+++ modules/backend/src/Au5.Application/Common/Abstractions/ITokenService.cs
@@ -3,4 +3,8 @@ namespace Au5.Application.Common.Abstractions;
 public interface ITokenService
 {
 	string GenerateToken(Guid extensionId, string fullName, string role);
+
+	Task BlacklistTokenAsync(string userId, string jti, DateTime expiry);
+
+	Task<bool> IsTokenBlacklistedAsync(string userId, string jti);
 }
