+++ modules/backend/src/Au5.BackEnd/appsettings.json
@@ -1,7 +1,7 @@
 ﻿{
   "ConnectionStrings": {
-    "ApplicationDbContext": "Server=.;Initial Catalog=Au5;Persist Security Info=False;Encrypt=False;TrustServerCertificate=False;Connection Timeout=30;"
-
+    "ApplicationDbContext": "Server=.;Initial Catalog=Au5;Persist Security Info=False;Encrypt=False;TrustServerCertificate=False;Connection Timeout=30;",
+    "Redis": "localhost:6379",
     // "ApplicationDbContext": "Server=tcp:au5.database.windows.net,1433;Initial Catalog=Au5Db;Persist Security Info=False;User ID=au5;Password=3cpTFq4Db+>WL7ev8-nJQE;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
   },
   "Logging": {
